+++ config/reader/src/localConfig.ts
